// packages/backend/src/modules/indicators/dto/indicator-filters.dto.d.ts
export declare class IndicatorFiltersDto {
    sportId?: number;
    regionId?: number;
    groupId?: number;
    search?: string;
    scope?: string;
    scopes?: string[];
    userId?: string;
    userOrganizationId?: number;
    userSportId?: number;
    ancestorOrgIds?: number[];
    userRole?: string;
}